Betty editing style warning corrected

+++ print_string.c
@@ -13,7 +13,7 @@ int print_string(char *input_str)
 
 	while (*input_str != '\0')
 	{
-		_putchar(*input_str);
+		_putchar((int)*input_str);
 		char_count = char_count + 1;
 		input_str++;
 	}
